Sign the user out and redirect home on auth failure (#88)
Users are sometimes seeing unexpected Ueberauth failures of unknown
provenance. Instead of sending a 403 unauthenticated response, we are
signing them out and sending them to the home page to start the auth
path over again.

We should be on the lookout for users getting trapped in a loop because
of this. If we observe that happening we should rethink this remedy.