Filter out buses whose block is over from the ladder
Criteria:
• No Swiftly assignment (the existing "is off course" status).
* Block finished more than an hour ago according to the schedule.

We give the buffer so that we don't tag a bus that is on a detour and
running late, and should thus still be shown on the route ladder.

We are preserving the vehicle in the data looking forward to a future
where you might be able to search for vehicles outside of the route
ladder context.