make minischedules data from hastus data
Preprocesses hastus data into the format we'll need for minischedules,
and stores it in Schedule.Data

This does not include a way to get it out of Data yet

Has a placeholder implementation for the details of turning trips
into pieces and understanding Operator and Sign-on activities.

The big picture of separating trips and activities into their blocks and
runs is done, and will let us start working on fetching this data into
the frontend, though the placeholder means the data won't be quite right.