Set up project for work on mini-schedules
Add a new mini-schedule laboratory feature flag.

Copy all properties panel-related files into a "STABLE" folder.
The files in the stable folder will remain unchanged as we develop mini-
schedule in the "non-stable" files. When we are ready to launch mini-
schedule, we will be able to delete the "stable" folder without having
to make any changes to the "non-stable" files.