Rework implementation of VPP/GPP tab controls
Several commits ago, we made the vehicle-status header at the top of the
VPP/GPP sticky via a simple CSS tweak. We would also like the tabs
themselves to stick, so that a user can scroll up and down the full
length of a run/block and not have to scroll back to the top to get the
controls back.

Unfortunately, the react-tabs library that we were using locked us into
a certain hierarchy of components, which was not compatible with how
we'd have to alter the page's structure to get the tabs in the sticky
container too. Here we get rid of the react-tabs library and roll our
own implementation of tabs.