Use normal scale on narrow screens (#764)
Use normal scale on narrow screens

The adjustment of scale that we remove here was meant to fit more content on narrow phone screens. However, it also meant that on iPhones, an ill-considered auto-zoom feature was kicking in unless we made the text unfeasibly big (24px vs. the usual 16px).