Clicking notification opens VPP
When a notification is clicked, the server tries to determine what
vehicle/ghost, if any, is related to the current trip out of the trips
affected by the notification. If there is such a vehicle or ghost, we
open the VPP with it, even if the ladder for the route in question is
not currently displayed.