Index vehicles by run ID for notification purposes
We had been indexing vehicles by trip ID in the realtime server, so that
when a user clicked a notification, we could use the trip IDs associated
with the notification to look up the pertinent vehicle.

This led to a counterintuitive behavior where, if a vehicle was on a run
affected by a block waiver, but not on one of the affected trips, it
would not be considered associated with the waiver.

Here we change to indexing and retrieving vehicles by run ID instead, to
avoid that problem.