fix: correctly handle conflicts on block_waivers (#1003)
Currently if there's a conflict on the block_waivers table it will
blow up completely instead of just rolling back the transaction. This
changes that, so that we can now guard against duplicate notifications
just with the unique index on block_waivers and not relying on the
unique index on notifications, which we're about to drop.