perf: pull pieces into a new table for better querying
The previous code needed to look through all of the pieces in a given
run to find out if the run was active in the given timeframe. This
commit pulls the pieces out into a separate table, which we can use to
only pull out the run IDs which match the given service ID/timeframe directly.