fix: remove typespec that was causing Credo error
The obvious thing to try would be changing the return type to
Notifications.Notification.t(), but the success typing is more narrow
than this because Dialyzer infers that certain keys will have specific
values. I tried various tricks to get around this but at the end of
the day didn't find a good solution, and the typespec here wasn't
particularly helpful.