chore: reimplement fullscreen control
The leaflet.fullscreen package includes its own implementation of
screenfull, but that ends up causing issues in tests and I can't
figure out how to properly get it to find its own implementation of
screenfull, or mock my own. As a result, I have added screenfull as a
separate dependency but for dev only to make the tests work.