fix(Map): Attatch zoomend event listener sooner
This resolves a race condition in the case where a map has 1 vehicle. The autocenterer was calling  before the zoomened listener was attached, so the zoom level was 16 but no garages were displayed. By attaching the zoomend listener earlier on in the EventAdder component, the zoomend event from  is received and the garages rendered as expected