fix(MapPage.test): Remove interdependency between tests
useVehicleForId was mocked with mockImplementationOnce in one test, but was never called as part of that test. That resulted in that mock being unexpectedly called during the subsequent test. This removes that stray useVehicleForId mock and also forces mocks to clear between tests in this file to prevent further interdependency.