feedback(design):style(css/_vehicle_route_summary): set `line-height` to `1`
looking back over the designs showed that the text was a lot closer to the top of the separator than I originally thought.
The `line-height` and `align-self: center` is currently preventing the text from reaching the top of the container.
So to make the text able to hug the top of the container we're setting the `line-height` to `1`