feat(nav): Updated search map title & icon (#1894)
* feat(Nav): New search map icon + page title

* fix(icon-search-map): Remove fill, height + width

* fix(IconSearchMap): Restore height/width and fix underlying viewbox issue

the height and width on the raw SVG file were both set to 24 (matching the viewbox definition). Surprisngly, the image was being cut off. I found that removing the height/width properties resolved the issue. Looking further into it, the viewport was being removed by svgo - see [this](https://github.com/svg/svgo/issues/1128) extensively discussed issue. Since we are re-scaling the svg with CSS to 20x20, we do need to preserve the image viewbox. Applying the solution from [this](https://github.com/svg/svgo/issues/1128#issuecomment-978136077) comment

* fix(GarageIcon): correct size

Going through all other icons to see which had a  viewbox that matched the width + height  and found only two: icon-garage and icon-up-right-arrow. icon-up-right-arrow had sizing explicitly set in CSS and was not affected. icon-garage defaulted to the leaflet default marker size of 12x12. This explicitly sets the iconSize to 21x25, matching what is specified in the SVG. I noticed that the width and height SVG properties are being removed as part of [svg-inline-loader](https://github.com/webpack-contrib/svg-inline-loader/tree/4bb5529fbdd847d7809067fe11840b48ee13dde4#removesvgtagattrs-boolean). I thought about turning off that behavior as part of this PR, but it seems like it could affect more icons and I'm not really sure if it is necessary

* test(SearchMapIcon): Add rendering test

* fix(SearchMapIcon): Adjust sizing/viewbox so not cut off