feat(ts/components/stopCard+mapMarkers+mapDisplay): keep stop card out from under the VPC/RPC area
On the Map Page (the only place the stop card is used right now), the VPC/RPC can occlude the Stop Card under certain conditions.
Things like screen size and where the stop is on screen affect where the stop card will open on screen.

This uses leaflet's existing padding functionality to keep the card on screen when clicked, by providing a padding value[^1] to leaflet
which includes the the left safe area as well as other controls we have on the screen by hardcoding these values for now.

A React Context was chosen to pass this information down the hierarchy due to the amount of code that
would need to be edited to allow for prop drilling down to the stop card. The stop card was also configured so that it
does not _require_ the context to function on it's own, but rather can take these arguments simply as props if desired,
which should help with unit testing and composition.

---

Asana Ticket: https://app.asana.com/0/inbox/1203014709797732/1203674636294159/1203947212316076

[^1]: https://leafletjs.com/reference.html#popup-autopanpaddingtopleft