feedback:refactor(css/_vehicle_map): move `StreetViewSwitch` `z-index` configuration into `c-vehicle-map`
Due to the `StreetViewSwitch` being an option on the map (`<Map allowStreetView/>`), I've moved the z-index configuration into the general map styles to colocate them. As long as `StreetViewSwitch` remains in the `<Map/>` I think this is correct, but if `StreetViewSwitch` is factored out in the future, this style should probably move elsewhere.