fix: Remove originalShapeClickable from DetourMap stories
In [this PR][1], we removed originalShapeClickable as an attribute for
the DetourMap, but in parallel, we introduced [storybook stories for
DetourMap][2], which referenced `originalShapeClickable` at the time.
Those two PR's were merged in rapid succession, so the builds for [the
second one][2] never caught up with the fact that [the first one][1]
had removed that argument.

[1]: https://github.com/mbta/skate/pull/2459
[2]: https://github.com/mbta/skate/pull/2457