Step 4, Part 1: Begin dropping or lifting useDetour tests
Now that the diversionPage tests are (mostly) restored, the code that
remains in useDetour is the code that's actually covered by the
diversionPage tests (mostly).

Now we start bringing the useDetour tests back, one-by-one, to see if
they should be dropped (deleted) or lifted (to the diversionPage
level). If they immediately pass, then they should (probably) be
dropped, although it's worth ensuring that the thing/s they test are
fully covered at the diversionPage level (rather than accidentally
being restored as a side effect of something else).

If they don't immediately pass, then that's either useful coverage
that should be moved to diversionPage (by adding a new test at the
diversionPage level that fails because of the same missing code as the
useDetour test), or testing an extraneous implementation detail (for
instance, a flag that diversionPage may not even be checking), which
means that both the detail and the test can be dropped.

Step 4, Part 1: Adding start and end points