fix(tests): Don't change the input on every useDetour test rerender
While debugging test issues with [this PR][1], I noticed that
useDetour.test.ts was re-invoking originalRouteFactory.build() every
time something would cause the hook to rerender, which can cause any
hooks that depend on the input to rerender as well (which is not what
would happen in real life, since the input isn't randomly changing).

This change updates `useDetourWithFakeRoutePattern` to take in the
`originalRoute` as an argument so that the input is only computed once.

[1]: https://github.com/mbta/skate/pull/2638