fix: Run database migrations before starting anything else
This allows us to put migrations and code that relies on those
migrations into the same commits.

If the code relies on changes made in a migration, and the migration
runs after the code starts, then there's a potential race condition
where that code runs before the migrations have run, which would cause
an error. This eliminates that possibility.