feat(ex/detours): add detour `:status` field, and populate from changeset (#2952)
* feat(ex/detours): add detour :status field, and populate from changeset

This adds a new database column and starts populating it from the changeset function.
Once we deploy this and have all instances become aware of the field. We'll then
deploy a data migration which adds this to previously made detours, so that there
are no running instances without this change to the changeset function

* chore(ex/db/detour): add type to field

* refactor(ex/detours): move `detour_type` into `Skate.Detours.Db.Detour`

* refactor(ex/detours): replace `detour_type` with `Detour.status`

* test(ex/detours): add tests for `:status` column

* feat(ex/db/detour): set `:status` in changeset when `:state` changes`