refactor(ex/notification_server): remove `:notify_finished`
This option was used only by tests as a way to be notified of broadcast
completion. I believe I completely overlooked that you could otherwise get this
information from messages sent to processes that `subscribe/2` to the server and
should have used that instead of implementing another message.

This removes `:notify_finished` references and code supporting it. All tests
that used `:notify_finished` were successfully updated to use `subscribe/2`
instead with 0 issues.