feat: broadcast detour expiration notification to users (#3016)
* feat(ex/notifications): add read interface for all notification types

* feat(ex/notification_server): implement generic notification pubsub

* refactor(ex/notification_server): remove `:notify_finished`

This option was used only by tests as a way to be notified of broadcast
completion. I believe I completely overlooked that you could otherwise get this
information from messages sent to processes that `subscribe/2` to the server and
should have used that instead of implementing another message.

This removes `:notify_finished` references and code supporting it. All tests
that used `:notify_finished` were successfully updated to use `subscribe/2`
instead with 0 issues.

* feat(ex/notification): send expiration notification to frontend

* feat(ex/notification_server): log pubsub activity

* refactor(ex/notification_server): remove detour specific broadcast