fix(ex/notification_server_test): fix flakey log test (#3022)
* fix(ex/notification_server_test): fix flaky log test

PR #3016 introduced new tests that were potentially not determinstic, but I had figured I'd addressed it. Turns out this is not the case https://github.com/mbta/skate/actions/runs/15884857906/job/44794195336 And I confirmed that the instability can be found with the following command
```sh
: , retry -d 0 --until=fail mix test ./test/notifications/notification_server_test.exs:359
```

So looking for a solution, I tried converting the `GenServer.cast` and `GenServer.abcast` calls to `call` and `multi_call`, but had a few difficulties, mainly that a GenServer calling itself with `multi_call` while handling a message in `handle_cast` locks up. This could be addressed in a few ways, one of them being to not have an intermediate `broadcast_to_cluster` message, but this seems to work and continue to use `cast`, since the result of the broadcast operation likely doesn't need to block the caller.

---

`:sys.get_state` seems like the best option because it uses some of the primary mechanisms of BEAM and OTP to avoid needing `Process.sleep`.

> Another option that’s more reliable than sleep is something like :sys.get_state - the target GenServer can’t handle the system message until handle_cast returns, so the test always waits exactly long enough.
>
> &mdash; https://elixirforum.com/t/logger-messages-not-caught-by-exunit/65064/3

So far, I have run this for 5-10 minutes (using `retry`) without fail. all other tests I've done have at least failed by this point. This is hundreds of tests at this point.

When I put `:sys.get_state` before the `assert_receive` it can still fail. If you pipe it to `dbg/1` it takes long enough to work, but it's not reliable without `dbg/1`, so putting this after makes sure we essentially do "SYN-ACK" with the Notification Server's `GenServer` to ensure it's done processing our request, without needing to convert the call to a synchronous `call` and we can keep the cast semantics.